+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog
 
 in progress
 ===========
-- Basic implementation generating UUIDv4 identifiers.
-- Add ULID identifiers.
+- Basic implementation generating UUIDv4 identifiers
+- Add ULID identifiers
+- Add ``--upper`` and ``--lower`` options for output transformation
 
 
 2019-06-28 0.0.0